refactor(hooks): split hook health module
Move hook health types, inspection, repair planning, prompts, warnings, and Codex trust helpers into focused submodules while preserving the existing crate::hook_health API.

Refs #332
